+++ shell.nix
@@ -1,0 +1,1 @@
+{ compiler ? "ghc8107" }@args: (import ./default.nix args).simpleSurvey.env