+++ src/SimpleSurvey.hs
@@ -1,0 +1,4 @@
+module SimpleSurvey where
+
+main :: IO ()
+main = print "Hello, world!"