+++ ChangeLog.md
@@ -1,3 +1,5 @@
-# Changelog for simple-survey
+# Revision history for simpleSurvey
 
-## Unreleased changes
+## 0.1.0.0 -- YYYY-mm-dd
+
+* First version. Released on an unsuspecting world.